output_files/crewai/instagram_post/pyproject.toml
[project]
name = "copy"
version = "0.1.0"
description = ""
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
requires-python = ">=3.10,<=3.13"

dependencies = [
    "crewai[tools]>=0.152.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.1",
    "langchain-community>=0.3.0",
    # TODO: add PyPI package for custom tool "BrowserToolsscrapeandsummarizewebsite"
    # TODO: add PyPI package for custom tool "SearchToolssearchinternet"
    # TODO: add PyPI package for custom tool "SearchToolssearchinstagram"
]

[project.scripts]
copy = "main:run"
train      = "main:train"

[tool.uv]
dev-dependencies = []

[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"
